refactor: Move hooks directory to src root
API fetch hooks can be shared between multiple tabs, therefore it
should be kept in the upper scope.
