+++ packages/frontend/src/hooks/useRoutines.ts
@@ -1,7 +1,7 @@
 import { useQuery, UseQueryResult } from "react-query";
 
-import { getRoutines } from "../services";
-import { IRoutine } from "../types";
+import { getRoutines } from "../Daily/services";
+import { IRoutine } from "../Daily/types";
 
 export function useRoutines(): UseQueryResult<IRoutine[], Error> {
   return useQuery<IRoutine[], Error>(`routines`, () => getRoutines());
