+++ packages/frontend/src/hooks/useWorkouts.ts
@@ -1,9 +1,9 @@
 import { useQuery, UseQueryResult } from "react-query";
 import { useRecoilValue } from "recoil";
 
-import { getDateString,getWorkouts } from "../services";
-import { dateState } from "../states";
-import { IWorkout } from "../types";
+import { getDateString, getWorkouts } from "../Daily/services";
+import { dateState } from "../Daily/states";
+import { IWorkout } from "../Daily/types";
 
 export function useWorkouts(): UseQueryResult<IWorkout[], Error> {
   const date = useRecoilValue(dateState);
